[keymgr_dpe/rtl] Fix `invalid_max_boot_stage`
Boot stage can be at most one less than the number of supported boot
stages (`DpeNumBootStages` synthesis parameter) for a successful Advance
operation, and boot stage is invalid for all values *greater than* or
equal to that maximum.

This inaccuracy in the RTL code did not lead to a problem in the current
instantiation, where `DpeNumBootStages = 4` and thus
`DpeNumBootStageWidth = 2`, where the width ensures that the value of
boot stage cannot be greater than `DpeNumBootStages`.  But this does not
hold generally and could cause problems if `DpeNumBootStages` is set to
a different value in the future.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>