[edn/dv] Turn off assertions when disabling and reenabling EDN in err vseq
This commit fixes some rare instances where the edn_err_vseq fails because the
data is unstable after the EDN has been disabled. The assertions can be disabled
here, since the specification says that the CSRNG and ES have to be disabled
and re-enabled before the EDN is turned back on.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>