[edn/dv] Add checks to scoreboard for the new cmd_ack and cmd_sts fields
This commit adds new checks to the scoreboard for the new cmd_ack and cmd_sts
fields of the sw_cmd_sts register. The cmd_rdy and cmd_reg_rdy fields are not
checked yet since there is another HW change planned where we remove the output
FIFO. It would be fairly complicated to track those two fields and since we plan
on changing the behaviour for those two regs a bit it would not be worth to write
the tracking code in the scoreboard yet.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>