[otbn,dv] Wait longer for an alert to come through in scoreboard
This timeout needs easing for the situation where we try to cause an
alert (by setting the LC escalation signal) near the start of time.
This will only be reported once the secure wipe is complete, but that
takes a while!

The previous wait (400 cycles) is not enough for the integrated
layout.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>