[rom/e2e] fix exit_failure matcher for bad policy / rollback tests
Currently for the bad policy and rollback tests, the exit failure matcher
will match BFVs while the success matcher will match two consecutive BFVs.
This works previously because we do a bulk read from UART and it contains
multiple lines, and success is matched before the failure. This is fragile.

This commit makes the matcher exclusive, so it doesn't matter how the read
is fragmented.

Signed-off-by: Gary Guo <gary.guo@lowrisc.org>