[keymgr_dpe, dv] Adds erase and disable to smoke and base seq
In base sequence adds erase and disable tasks that perform both
operations along with random advance and generations. Adds checking
enchancements as well for error scenarios.

In smoke sequence re-works it so that it performs disable operations
at the beginning and end of the test. The first time being in reset and finally
while in available state. It fixes the comment structure to better explain
the flow of the smoke test, and adds a few more stressfull scenarios like
advances with key_slots all full and a advance with retain_parent == 0.

Signed-off-by: Bobby Reyes <bobby@rivosinc.com>