[keymgr_dpe, dv] Adds SCB handling for disable operation
Adds the handling of disble operation in the scoreboard.

When entering into disable state checks that internal key's valid
is de-asserted and any further operations trigger an error.

The sideload keys are also checked. From any previous successfull
genrate the exp key is updated in the keymgr_dpe inteface. Upon reset
these values are compared against the current sideload key values.
An error is thrown if they are not equal

Signed-off-by: Bobby Reyes <bobby@rivosinc.com>