[otp_ctrl,dv] Fix types in otp_ctrl_scoreboard.sv
This fixes a silly typo in the definition of a "check_timout"
variable. In particular, simulators like Xcelium don't allow a
variable to be declared like "bit [123] foo". The correct syntax would
be packed dimensions like "bit [123:0] foo".

However, there's a reasonably obvious simpler fix, so I switched
things to just using an integer. In the original code, a value of 0
was overridden with '1 (which was 2**32 with this type). Leaving it as
zero (which already means "no timeout" for set_exp_alert) seems to
work perfectly well.

As a minor nit, I also changed the variable name. Timeout now contains
an "e"!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>