[chip-level/sival] Set two chip level EDN tests timout to long
This commit sets the time out for two of the chip level EDN tests
to long. These tests were failing or are now failing with the added
HW changes, because the timeout values were too low. This commit
fixes this issue and removes the broken tag from one of the tests.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>