[edn/dif] Align dif with new sw_cmd_sts register
This commit aligns the EDN difs with the HW changes in prior commits.
Some of the difs are shared between the EDN and CSRNG. The main difference
is that now the difs can see when a command is done by looking at the
ack bit in the sw_cmd_sts register.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>