[bazel] Introduce a stamping module
The purpose of this module is to enabled rules to detect whether
stamping is enabled on the command-line or not. It requires some
hackery inspired by a bazel bug thread and which is very similar
to how rules_rust handles this as well.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>
(cherry picked from commit bc4fd072abebf52b1bf15bb540115c5c245f183a)