[bazel] Convert `autogen_chip_info_src` to stamping
With this change, `autogen_chip_info_src` will only pass version
information if stamping is enabled on the command-line. This
commit changes the meaning of the `--ot_version_file` flag of
`//util:rom_chip_info`. Previously, it was a path to a file
that contains the git sha. It is now a path to the bazel version
info file which is parsed using the `version_info` library.

Note: with this change, unless stamping is enabled, the ROM will
not contained the git sha in the chip_info anymore.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>
(cherry picked from commit 375a0379cf5cf1f7bd073359332157c60758ecec)