[rtl,pwrmgr] Fix escalation timeout persistence
The escalation timeout needs to persist until reset. The timeout is detected by
clk_i, but the timeout is used to drive a flop clocked by clk_lc. This change
adds a synchronizer from clk_i to clk_lc, and a clk_lc flop reset by rst_lc_n.
The output of this flop is used to set the fault_status and as a reset request.

With this fix the pwrmgr_escalation_reset test passes at the block level.

Fixes #20516

Signed-off-by: Guillermo Maturana <maturana@opentitan.org>