[rv_dm,dv] Add comments to rv_dm_jtag_dmi_debug_disabled_vseq
The vseq was a reasonably magic bunch of instructions and it wasn't
particularly clear to me what it was supposed to be doing. This commit
adds comments about what's going on (and why we expect particular
values).

I've also slightly strengthened the test:

 - Randomise the arbitrary data instead of hard-coding something.

 - Read back abstractdata values and check they change in the expected
   way.

 - Make the timing waits a bit more general.

Note that the wait code was previously broken because it was calling
$urandom_range with backwards arguments. The wait would always have
been for zero cycles.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>