[prim] Avoid unnecessary Impl parameter in prim_onehot_check
This use of Impl was a workaround that I believe only really affected
some RTL that isn't in the repository(!). Get rid of the workaround:
we might be able to do it more robustly if we see the problem again.

Also add a bit of documentation comment about what's going on with the
assertion the Impl parameter was gating. It took me some time to
understand what it was trying to achieve.

As a bit of background: I only noticed this code at all because its
use of the parameter keyword isn't strictly compliant with the
SystemVerilog spec. I was about to convert it to using localparam,
when I was convinced that the usage never could have worked anyway. So
I thought it made more sense to rip it out!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>