[opentitantool] Introduce JSON config for teacup shield
The HyperDebug "shield" has a number of MUX'es, buffers and other
ICs (current sensors), which can be controlled through some signals in
addition to what is used when e.g. simulating OpenTitan on an FPGA.

This PR introduces a new `hyperdebug_teacup.json` file, to be used with
`--interface=hyperdebug` (not `hyper310`).  Both this file and existing
FPGA-specific files refer to a common `opentitan.json`.

Change-Id: I15390cd3ac93a149ea65635c3c3f298e7a08cea0
Signed-off-by: Jes B. Klinke <jbk@chromium.org>