[sival,rstmgr] Adjust rstmgr_alert_info_test
A minor cleanup of the test to expect an otp_ctrl error if not
in rom_ext. There is still an unexpected flash_ctrl read error
for cw310 upon entry to test_main, so it is pending from rom_ext.

Add utility to log the bits set in flash_ctrl fault status.

The issue #20589 has an explanation of this test's failures.

Signed-off-by: Guillermo Maturana <maturana@opentitan.org>