[util/reggen] Do not generate stamping information if not available
If stamping is disable, version_stamp will be an empty dictionary
and the code will print something like
```
// Generated register constants for rv_plic.
// Built for <unknown>
// https://github.com/lowRISC/opentitan/tree/<unknown>
// Tree status: <unknown>
// Build date: <current date>
```
This is unhelpful and still prints the current time which is defeats
the point of not stamping. The new code will not print anything
when a piece of information is not available.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>