[rv_dm,dv] Remove a TODO in rv_dm_env.core
The TODO was added a long time ago because the environment doesn't
actually need the whole of an rv_dm implementation. But the core file
on which we depend is just a "grab the vendored rv_dm repository".
It's probably not really worth putting in engineering effort to chop
it up: I don't think we'll ever want to "vendor just some of the
repository".

Drop the TODO message.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>