[rv_dm] Fix an assertion
The assertion is for the same condition as in the
pinmux_strap_sampling module. However, as opposed to that module,
there are some modifications to the strap_en_i signal resulting in
a local version of it (strap_en) and the SVA needs to be
formulated for that one instead.

Signed-off-by: Michael Schaffner <msf@opentitan.org>