[RV_DM]rv_dm_scoreboard
This commit contains a change in line of code that is passing criteria for our 'rv_dm_sba_debug_disabled_vseq'. Our test condition is that no transaction was seen on TL SBA interface when 'lc_hw_debug_en' is Off.So that's why we add this line in scoreboard.

Signed-off-by: Shahid Mehmood <shahid.impare@gmail.com>