[edn/rtl] Handle backpressure from CSRNG without an output FIFO
This commit enables the EDN to handle backpressure from CSRNG by introducing
handshakes and gating the propagation of words of the commands. Words now only
move down the pipeline when CSRNG is ready to accept the next word.

To enable this change, this commit introduces some tweaks to the main SM.
The main SM now has some new states in which the EDN is notified that
the corresponding boot commands can be sent.

This commit also handles the fields of the sw_cmd_sts accordingly.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>