[rom] Revert post-ES ROM changes
Restore the binary ABI between the ROM and the ROM_EXT to what was
released in the ES chip.

This partially reverts commits:
- cf62de118db
- b75df1e0f56
- 6905bbf44f0

Along with commit 97ebef0bbef, this should undo all binary incompatible
changes to the ROM since the ES tapeout.  This is for the
earlgrey_es_sival branch only to create a more accurate ROM-to-ROM_EXT
handoff on this branch.

**Do not cherry-pick this to master.**

Signed-off-by: Chris Frantz <cfrantz@google.com>