[sca] Fix trigger parsing
This PR fixes a bug in how the trigger select is parsed. The ot-sca framework
sends a 0 to select the HW or 1 to select the SW trigger. To parse this bit,
we need to  use the read_32() function.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>
(cherry picked from commit 78903f71cb03e41b45108dd2fb943835ec919e00)