[rom_ext] Improve error reporting for bad images
Trying to boot a bad image often results in
`kErrorBootPolicyBadIdentifier` rather than the true error code for the
bad image.  This is because the ROM_EXT tries the non-primary slot which
is often empty and the "bad identifier" error is appropriate for an
empty slot.

1. Determine if the error is "bad identify" for the non-primary slot and
   prefer the primary slot error message in that case.
2. Link the customer signing keys into the "fake signed" ROM_EXT targets
   to facilitate creating ROM_EXTs for FPGAs or DEV parts.

Signed-off-by: Chris Frantz <cfrantz@google.com>
(cherry picked from commit 045130705360dafe3dfec18c313797abe9077736)