[cryptolib] Add driver code for P-384 ECDH and ECDSA on OTBN
This commit adds Ibex driver code for P-384 ECC on OTBN. In case of
ECDSA the code is split up into individual files for keygen, sign, and
verify. This is because each function is a separate OTBN application.
But still, everything is put together in a single ecdsa_p384 file, which
then is used by the top level ecc driver file.

Similar, curve point validation is put into an own driver file, as it
also is a separate OTBN application. It is executed before ECDH shared
key generation and ECDSA signature verification, to check if the public
key is valid (= is a curve point).

Co-authored-by: Bojan Uljarevic <bojan.uljarevic@gi-de.com>
Signed-off-by: Moritz Wettermann <moritz.wettermann@gi-de.com>