[opentitantool] HyperDebug support for baud rate and break condition
HyperDebug uses a non-standard bulk USB protocol for serial ports, and
the Linux kernel has only a rudimentary driver, meaning that the usual
system calls to set parity and baud rate or generate break conditions,
do not work.

This CL adds code to opentitantool to be able to issue the USB control
requests directly, instead of going to the Rust serial port library.

Required change to Hyperdebug firmware:
https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5241764

Change-Id: I654e8fbc5f88d2c4e57dcaf5229047cf71e8b810
Signed-off-by: Jes B. Klinke <jbk@chromium.org>