Revert "[edn] Move prim_edn_req out of prim"
This reverts commit 3b4e36e01c130d75844efe5d5b2a31502059e74c. The
reason is that this will change some paths in the hierarchy and we're
worried that this will cause slightly pointless work for someone with
CDC/RDC waivers.

Honestly, it seems a bit rubbish that this is something we need to
worry about! But we can make things self-contained in a less intrusive
way, so that's probably the right approach.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>