[sca] Add context for SW LFSR
At the moment, a 32-bit SW LFSR is used for masking data shares
sent from IBEX to crypto-blocks (AES, KMAC, SHA3).
This commit adds a new context to this PRNG so that it can be used
for multiple purposes. This is needed to set the stage for
implementing aes-fvsr-data capture which makes use of this LFSR
to determine the order of measurements.
This commit also ammends all files that make use of the lfsr
functions.

Signed-off-by: Vladimir Rozic <vrozic@lowrisc.org>
(cherry picked from commit f75d8945d01f163592df2ce1cef70a4099643fb2)