[dv] Correct direct prediction of regwen
Direct prediction should alter the mirrored state no matter
the access policies. However, this was not the case before,
which means that if the scoreboard changed a 0 regwen back to
1 with direct prediction, the lock-states of the affected CSRs
were not updated correctly. This prediction mode is useful when
the regwen can be locked both by HW and SW.

Signed-off-by: Michael Schaffner <msf@opentitan.org>