[spi_device] Enable continuously polled status registers
Previously, the Read Status Register commands behaved differently from
the SPI flash convention, holding a fixed value for every word clocked
out of the same command + rotating through all three registers for each
command. Change the commands to only return the byte targeted by the
specific command, and update the status register value at every byte
beat (8th posedge of SCK).

This brings the SPI device flash behavior in line with the convention,
which is to allow hosts to poll the same register repeatedly and see
updated values as it goes along.

Disable the scoreboard's validation of Read Status commands for now. It
uses csr_rd() to determine the resolved value, but this is impossible.
The source of truth is in the SPI domain, and it updates at a much more
frequent interval than the core clock domain: The SPI domain updates for
every 8th posedge of SCK, whereas the core clock domain's view updates
only at every CSB de-assertion (delayed by synchronization to an async
domain).

In addition, add a CSR to permit resetting the async FIFO used to bring
software-originated status register changes into the SPI clock domain.

Signed-off-by: Alexander Williams <awill@opentitan.org>