[crypto] Precompute CRC32 checksums for OTBN apps.
This adds code to `otbn_build.py` that computes the checksum and stores
it as a symbol in the binary, similar to how we store the start address
of the `.data` section. When Ibex loads the binary into OTBN, it now
checks that the `LOAD_CHECKSUM` register matches the precomputed value
after the app is loaded.

Signed-off-by: Jade Philipoom <jadep@google.com>