[spi_device] Write mode and status bits alongside command
Write some pertinent mode and status bits to the cmdfifo alongside the
command. The BUSY and WEL bits can be used to determine whether an
uploaded command should be accepted, and the address mode bit helps
determine whether an associated address in the addrfifo should be
interpreted as 3 bytes or 4 bytes.

Uploading these bits alongside the command byte avoids race conditions
with software checking their values after the transaction actually
occurs.

Make WEL an explicit field in the FLASH_STATUS CSR, and mark it as only
RW0C. Like the BUSY bit, software has no need to set this field to 1. It
should only clear it when it completes uploaded commands. Making it RW0C
eliminates the need for read-modify-write cycles just for checking
hardware-writable bits.

Signed-off-by: Alexander Williams <awill@opentitan.org>