[otp_ctrl] Differentiate between owner and creator keys
This updates the templating mechanism so that keymaterial
can be associated either with the creator or the owner.

This is needed since the life cycle controller outputs
different enablement signals for creator and owner seeds.

Note that because key material is stored in secret partitions,
the key material will be inaccessible by SW after locking the
associated partition.

Outputing the key material to the key manager requires all
partitions holding keymaterial to be locked.

This addresses #21044.

Signed-off-by: Michael Schaffner <msf@opentitan.org>