[sca] Prepare SHA3 for pen. tests
Currently, the SCA code for SHA3 uses features to improve SCA
that are only available on the FPGA but not on the discrete chip.

This PR, similar to #20772, changes the sha3_sca code such that
the code supports measurements on the FPGA and on the chip.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>
(cherry picked from commit f7586cc9e8081fdd50a914d628de9121a9fc22e2)