[keymgr_dpe, dv] fix for keymgr_dpe smoke failure
Makes csr_wr for erase non-blocking. In some cases it seemed that the check for wait_op_done
was happening after the operation had actually concluded and the dst slot was erased. Thus
a false failure was occuring as for erase operations the sequence ensures that a empty key slot
cannot be erased. The csr_wr being blocking was causing this delay in those cases.

Signed-off-by: Bobby Reyes <bobby@rivosinc.com>