[edn/dv] Fix newly failing edn_err vseq
This commit fixes the failing edn_err vseq. Don't exactly know why
it didn't fail before but this commit should fix it. The issue was
that disabling the EDN clers the auto mode FIFOs. This means the FIFOs
need to be refilled before re-starting the EDN, otherwise the auto mode
does not have any commands to take out of the FIFOs.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>