[edn/dv] Adapt DV to the new changes
This commit adapts the DV to the new changes. Now that the CSRNG
sts error signal causes the EDN to stop accepting entropy, it can
not be random anymore. Furthermore the err_vseq causes the EDN to
enter an error state, for which there is no guarantee that the
valid signal stays high.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>