[rv_dm,dv] Add a short wait in rv_dm_smoke_vseq
This avoids a problem with the rv_dm_tap_fsm_rand_reset sequence. That
sequence runs a smoke vseq and then tries to reset it in the middle of
its operation. That might not work if the reset time happens to match
a big block of back-to-back CSR accesses.

The change in this commit ensures that there are always occasional
cycles where there isn't a CSR access going, which means the reset
sequence can safely inject a reset.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>