[opentitantool] Fix for SPI chip select and opentitansession
The JSON configuration files allows the declaration of aliases for SPI
and I2C ports, and further allows each alias to specify clock speed and
chip select (or I2C address) to use by default when that alias is used.

That feature will be used to declare which chip select to use for SPI
TPM communication (as the HyperDebug default goes to the OT SPI
passthrough chip select).

The TransportWrapper layer has code in `src/app/{spi,i2c}.rs` to keep
track of which alias was last used for any given physical SPI/I2C port
of the transport, so that the transport is instructed to change speed or
other settings only when a new alias is uses, and not when the same
alias is used for multiple subsequent operations.  This is a significant
optimization, but also affects correctness, since a SPI port may be
instructed to hold chip select asserted across several operations, and
the HyperDebug transport is not able to keep chip select asserted, if
told to switch chip select pin (even if told to use the same as last
time.)

It turns out that the way opentitansession receives commands via a
socket, carrying the name of a SPI port, it currently creates a new
`spi::Target` instance for each request, even if it the same name.  The
logic in `src/app/{spi,i2c}.rs` is not prepared for multiple `Rc<>`
instances referring to the same port alias, and treats them as distinct
aliases (causing HyperDebug to deassert chip select).

This CL makes sure that the `TransportWrapper` keeps track of
`spi::Target` instances already created, and makes it return the
existing instance if the alias has already been used.

As the Google Ti50 team uses `opentitansession` extensively in our
automated testing, this PR is instrumental to our ability to use the
Teacup board in our test lab.

Change-Id: I35ed00e2b29d47a41bf2711d33dae2df770566c0
Signed-off-by: Jes B. Klinke <jbk@chromium.org>
(cherry picked from commit 6de5099778f9dfbf2b4f65cb92dba40983a7407a)