[flash_ctrl/dv] Simplify cov interface and remove NoSameAddrRead_A
Previously, NoSameAddrRead_A would sometimes fire incorrectly since the
model of the read buffer is too simplistic in this interface. Modelling
the read buffer accurately would require access to more internal signals
from the design so that all corner cases (e.g. integrity errors) can be
modelled accurately.

The assertion is therefore removed, as it is not strictly required for
the coverage interface to fulfill its purpose.

Signed-off-by: Michael Schaffner <msf@opentitan.org>