[keymgr_dpe/doc] Fix diagram with derivation chain example
The previous diagram did not override the UDS (boot stage 0 key) when
deriving the creator key (boot stage 1 key), but the UDS *must* be
overridden (i.e., it cannot be retained).  This commit fixes that.

The diagram is stored on the OT gDrive at
https://drive.google.com/file/d/1mVJC3_WuoFSXHLQy4ygn75TUa9pNOkqp/view
and got exported with the settings listed in
7ce70543bc7184f1f3f78b0577b15e4fc4449054.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>