[spi_device] Fix up TPM command handling
Move the TPM Write and Read FIFOs to the SRAM. Change the Write FIFO's
interface to be more like the SPI flash upload type, where the hardware
side is a FIFO, but the software side is a RAM with an explicit
busy/free bit. As a side effect, the Write FIFO now has a wider
interface, and buffer reads can proceed more rapidly.

Adjust the synchronization mechanisms with software to eliminate
ambiguity in the relationship between data buffers and commands.

For TPM write commands, wait until the data is completely transferred
before pushing the command into the command FIFO. In addition, do not
pass the wait state if the command FIFO is not empty and if the write
FIFO has not been released back to hardware. Aborted write commands will
simply be dropped, and software does not get confusing information. The
process works like this:

0. Write command enters the SPI hardware.
1. HW waits for the command FIFO to be empty and for the payload buffer
   to be released.
2. HW proceeds past wait states until just before the last posedge for
   the last bit specified by the command's byte count (xfer_size_met).
3. HW writes the command to the command FIFO on that last posedge. The
   SRAM is also written here for the payload.
4. SW reads the command from the command FIFO and notes the length of
   the data. A write command being present signifies the full data is in
   the write FIFO.
5. SW reads the data from the write FIFO, then writes the TPM_STATUS CSR
   to release the buffer back to HW.

For TPM read commands, wait until the command FIFO is empty before
pushing a new read command into it. In addition, hold the Read FIFO in
reset on the sys_clk side until a new read command is popped from the
command FIFO. "New" means a command that was pushed since CSB was most
recently asserted. After this handshake, the Read FIFO's reset is lifted
and becomes open for writing. The process works like this:

0. RdFIFO is held in reset. It must be empty and can't be written yet.
   Read command enters the SPI hardware.
1. HW waits for SW to empty the command FIFO.
2. HW writes the read command to the command FIFO. This should happen
   basically instantly for software (relative to the command FIFO
   emptying).
4. Software reads the command FIFO to lift the RdFIFO reset. This gate
   prevents SW from writing stale data.
5. HW waits for the RdFIFO to have enough data.
6. HW completes the command.
7. HW puts the RdFIFO back in reset once CSB de-asserts.

In addition to those mechanisms, add a new interrupt to signal when data
was dropped from the RdFIFO. This includes both writes occurring while
the RdFIFO is held in reset and a read command not transferring its
proscribed number of bytes. It does not include data that software wrote
into the FIFO beyond what the command called for.

Signed-off-by: Alexander Williams <awill@opentitan.org>