[cw340] Relax SPI timing to 12.5 MHz
Vivado 2021.1 can choose poor placement for the SPI clock path for
passthrough, resulting in timing failures at 20 MHz. Relax the SPI
timing requirement to 12.5 MHz for now.

Location constraints (or a tool version upgrade) may be the better path
forward in the future, but with CW310's SPI constraints still at 20 MHz,
it can be the detector for timing issues for passthrough for now.

Signed-off-by: Alexander Williams <awill@opentitan.org>