[kmac] Add param to keccak_round/2share to only use external randomness
Depending on the PRNG architecture and control, the externally provided
randomness can be guaranteed to be stable when the inputs to the DOM
multipliers don't change. Not using partial intermediate results to
cover these cases allows saving some silicon area (minus 800 MUX2).

However, it seems that PROLEAD currently cannot successfully analyze
the design with this new option enabled. For this reason, we keep the
multiplexers in the design.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>