[dma/dv] Remove `dma_illegal_transfer` testpoint from testplan
The way this testpoint was described, any direct transfers between CTN
and SYS would be illegal.  This does not correspond to the current
specification (nor the implementation) of the DMA, however.  We have
thus decided to remove this testpoint.

The `dma_illegal_transfer` test that was listed as covering this
testpoint does not exist.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>