[rom_ctrl,dv] Minor tidy-up in rom_ctrl vseqs
The main change here was in rom_ctrl_corrupt_sig_fatal_chk_vseq, where
I've squashed the contents of the chk_fsm_state task into the
check_for_alert task. This means that we can combine the explicit
calls to the two tasks that appeared in that vseq.

I suspect that the original author intended to do the FSM state check
every time, but the code didn't quite do that. In this commit, I'm
using a check_fsm_state argument in order to reproduce the behaviour
that we have.

It turns out that the contents of this task were also inlined in
rom_ctrl_kmac_err_chk_vseq. To get rid of the duplication, I've moved
the task into the base vseq and used it in both places.

There are some small differences (number of wait cycles in two
places). I've passed them as an argument to the task and made the
default be the value that was used at most of the call sites.

To make things a bit clearer, I've also been explicit on how
the (default) check=UVM_CHECK argument to csr_rd makes the call to
uvm_reg::predict() do something.

Finally, that return value from predict() gets cast to void. This
return value will be 1 to say that the prediction did something and
Xcelium produces warnings about an ignored return value if there's no
cast.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>